:art: Rename schema
